Allow editing of gdocs in markdown with pandoc
vim-metarw-gdrive retrieves documents from google docs by looking for a
"downloadUrl" field in the api response.  This field will not be present
if the doc was created using google docs itself (a gdoc).  In order to
edit these simple documents, this patch requests google to export the
doc as html.  Back inside the plugin, this patch uses pandoc to convert
it to markdown.  The process is reversed when the file is saved from
within vim.